+++ src/ext/backtrace.h
@@ -1,0 +1,7 @@
+#ifndef DD_BACKTRACE_H
+#define DD_BACKTRACE_H
+#define MAX_STACK_SIZE 1024
+void ddtrace_backtrace_handler(int sig);
+void ddtrace_install_backtrace_handler();
+
+#endif  // DD_BACKTRACE_H